+++ src/CustomizeGeyser/Changelog.md
@@ -1,10 +1,9 @@
 # Changelog
 
-## [1.0.6.0]
+## [1.0.6.0] EX1 S11-464102
 
 ### Fixed
 - added missing slush_salt_water to the default list
-- fixed geysers being random even after RandomizerEnabled set to false; note that since the DLC a lot more geysers are predetermined, so randomizing only fully works on world generation!
 
 ## [1.0.5.0] EX1 S6-448916